pyproject.toml
[tool.poetry]
name = "14-parser-fgis-arshin"
version = "0.1.0"
description = ""
authors = ["Alexey Inkov"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.32.3"
uvicorn = "^0.30.1"
fastapi = "^0.111.0"
loguru = "^0.7.2"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.31"}
aiosqlite = "^0.20.0"
kafka-python = "^2.0.2"
djangorestframework = "^3.15.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

